Moves sidebar component to its own directory
