+++ src/pages/contact.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { graphql } from 'gatsby';
 import Helmet from 'react-helmet';
 import Layout from '../components/layout/layout';
-import Sidebar from '../components/layout/sidebar';
+import Sidebar from '../components/sidebar/sidebar';
 import ContactForm from '../components/contactForm/contactForm';
 
 export default class ContactPage extends React.Component {
